+++ NEWS.md
@@ -1,3 +1,19 @@
+# ReboundTools 0.1.5 (2021-01-08)
+
+* Now at 303 tests, all passing.
+* Test coverage is 99.84 %. 
+* New graph parameters:
+    - `linend`
+    - `linejoin`
+    - `include_start_point`
+    - `start_point_size`
+    - `start_point_shape`
+    - `include_end_arrow`
+    - `arrow_angle`
+    - `arrow_length`
+    - `arrow_type`
+
+
 # ReboundTools 0.1.4 (2021-01-07)
 
 * Added x and y labels for graphs, when possible.
@@ -15,6 +31,7 @@
 * Package test coverage is 99%
 * Eliminated all errors and warnings.
 * Added documentation for functions.
+
 
 # ReboundTools 0.1.2 (2021-01-03)